fix: bundleImpl
