+++ packages/module-federation-metro/src/commands/bundle-remote/index.ts
@@ -8,10 +8,11 @@ import type { RequestOptions, OutputOptions } from "metro/src/shared/types";
 import type { ModuleFederationConfigNormalized } from "../../types";
 import { CLIError } from "../../utils/errors";
 import { Config } from "../types";
-import loadMetroConfig from "../utils/loadMetroConfig";
-import relativizeSerializedMap from "../utils/relativizeSerializedMap";
 import { createResolver } from "../utils/createResolver";
 import { createModulePathRemapper } from "../utils/createModulePathRemapper";
+import loadMetroConfig from "../utils/loadMetroConfig";
+import { saveBundleAndMap } from "../utils/saveBundleAndMap";
+
 import { BundleFederatedRemoteArgs } from "./types";
 
 const DEFAULT_OUTPUT = "dist";
@@ -45,46 +46,6 @@ async function buildBundle(server: Server, requestOpts: BundleRequestOptions) {
   });
 
   return bundle;
-}
-
-async function saveBundleAndMap(
-  bundle: { code: string; map: string },
-  options: OutputOptions,
-  log: (msg: string) => void
-) {
-  const {
-    bundleOutput,
-    bundleEncoding: encoding,
-    sourcemapOutput,
-    sourcemapSourcesRoot,
-  } = options;
-
-  const writeFns = [];
-
-  writeFns.push(async () => {
-    log(`Writing bundle output to:\n${chalk.dim(bundleOutput)}`);
-    await fs.writeFile(bundleOutput, bundle.code, encoding);
-    log("Done writing bundle output");
-  });
-
-  if (sourcemapOutput) {
-    let { map } = bundle;
-    if (sourcemapSourcesRoot != null) {
-      log("Start relativating source map");
-
-      map = relativizeSerializedMap(map, sourcemapSourcesRoot);
-      log("Finished relativating");
-    }
-
-    writeFns.push(async () => {
-      log(`Writing sourcemap output to:\n${chalk.dim(sourcemapOutput)}`);
-      await fs.writeFile(sourcemapOutput, map);
-      log("Done writing sourcemap output");
-    });
-  }
-
-  // Wait until everything is written to disk.
-  await Promise.all(writeFns.map((cb) => cb()));
 }
 
 function getRequestOpts(
